GradientBoostRegressor
Attached is the completed code for my gradient boost regressor. I trained it many times with many hyperparameters to get those that perform generally the best given the supplied base train/test data. The .csv contains my results for hyperparameters and the 'best_loss' they achieved, and at the bottom is the defined the single best GradientBoostRegressor and the hyperparams which achieve this. Any questions please ask!